requirements.txt
XenAPI
pyyaml
mysql-connector-python
